+++ Core/FormContent.cs
@@ -280,7 +280,7 @@ namespace Greatbone.Core
             return this; // ignore ir
         }
 
-        public FormContent Put(string name, IData v, byte flags = 0)
+        public FormContent Put(string name, IData v, ushort flags = 0)
         {
             if (name != null)
             {
@@ -474,7 +474,7 @@ namespace Greatbone.Core
         }
 
 
-        public FormContent Put<D>(string name, D[] v, byte flags = 0) where D : IData
+        public FormContent Put<D>(string name, D[] v, ushort flags = 0) where D : IData
         {
             if (name != null)
             {
@@ -500,7 +500,7 @@ namespace Greatbone.Core
             return this;
         }
 
-        public FormContent Put<D>(string name, List<D> v, byte flags = 0) where D : IData
+        public FormContent Put<D>(string name, List<D> v, ushort flags = 0) where D : IData
         {
             if (name != null)
             {
