+++ Core/XElem.cs
@@ -216,7 +216,7 @@ namespace Greatbone.Core
             throw new NotImplementedException();
         }
 
-        public bool Get<D>(string name, ref D v, byte flags = 0) where D : IData, new()
+        public bool Get<D>(string name, ref D v, ushort flags = 0) where D : IData, new()
         {
             throw new NotImplementedException();
         }
@@ -251,17 +251,17 @@ namespace Greatbone.Core
             throw new NotImplementedException();
         }
 
-        public bool Get<D>(string name, ref D[] v, byte flags = 0) where D : IData, new()
+        public bool Get<D>(string name, ref D[] v, ushort flags = 0) where D : IData, new()
         {
             throw new NotImplementedException();
         }
 
-        public bool Get<D>(string name, ref List<D> v, byte flags = 0) where D : IData, new()
+        public bool Get<D>(string name, ref List<D> v, ushort flags = 0) where D : IData, new()
         {
             throw new NotImplementedException();
         }
 
-        public D ToObject<D>(byte flags = 0) where D : IData, new()
+        public D ToObject<D>(ushort flags = 0) where D : IData, new()
         {
             throw new NotImplementedException();
         }
@@ -283,12 +283,12 @@ namespace Greatbone.Core
             throw new NotImplementedException();
         }
 
-        public D[] ToArray<D>(byte flags = 0) where D : IData, new()
+        public D[] ToArray<D>(ushort flags = 0) where D : IData, new()
         {
             throw new NotImplementedException();
         }
 
-        public List<D> ToList<D>(byte flags = 0) where D : IData, new()
+        public List<D> ToList<D>(ushort flags = 0) where D : IData, new()
         {
             throw new NotImplementedException();
         }
