+++ Sample/Message.cs
@@ -15,7 +15,7 @@ namespace Greatbone.Sample
 
         internal DateTime time;
 
-        public void ReadData(IDataInput i, byte flags = 0)
+        public void ReadData(IDataInput i, ushort flags = 0)
         {
             i.Get(nameof(fromid), ref fromid);
             i.Get(nameof(from), ref from);
@@ -24,7 +24,7 @@ namespace Greatbone.Sample
             i.Get(nameof(time), ref time);
         }
 
-        public void WriteData<R>(IDataOutput<R> o, byte flags = 0) where R : IDataOutput<R>
+        public void WriteData<R>(IDataOutput<R> o, ushort flags = 0) where R : IDataOutput<R>
         {
             o.Put(nameof(fromid), fromid);
             o.Put(nameof(from), from);
